Consolidate percentile from neighbourhood unit tests
